move SaveManifest to info.go
