+++ snappy/info.go
@@ -21,12 +21,16 @@ package snappy
 
 import (
 	"fmt"
+	"os"
 	"path/filepath"
 	"strings"
+
+	"gopkg.in/yaml.v2"
 
 	"github.com/ubuntu-core/snappy/dirs"
 	"github.com/ubuntu-core/snappy/progress"
 	"github.com/ubuntu-core/snappy/snap"
+	"github.com/ubuntu-core/snappy/osutil"
 )
 
 const (
@@ -187,4 +191,36 @@ func PackageNameActive(name string) bool {
 // ManifestPath returns the would be path for the snap manifest.
 func ManifestPath(s *snap.Info) string {
 	return filepath.Join(dirs.SnapMetaDir, fmt.Sprintf("%s_%s.manifest", s.Name, s.Version))
+}
+
+/// XXX: temporary step until we know if we still need this (vs overlord state)
+// and to find out what we absolutely need from it
+type diskManifest struct {
+	// XXX likely we want also name and snap-id and summary
+	Revision    int    `yaml:"revision"`
+	Channel     string `yaml:"channel"`
+	Developer   string `yaml:"developer"`
+	Description string `yaml:"description"`
+}
+
+// SaveManifest saves the manifest at the designated location for the snap containing information not in the snap.yaml.
+func SaveManifest(rsnap *snap.Info) error {
+	m := &diskManifest{
+		Revision:  rsnap.Revision,
+		Channel:   rsnap.Channel,
+		Developer: rsnap.Developer,
+		// XXX capture also Summary?
+		Description: rsnap.Description,
+	}
+	content, err := yaml.Marshal(m)
+	if err != nil {
+		return err
+	}
+
+	if err := os.MkdirAll(dirs.SnapMetaDir, 0755); err != nil {
+		return err
+	}
+
+	// don't worry about previous contents
+	return osutil.AtomicWriteFile(ManifestPath(rsnap), content, 0644, 0)
 }
